Split resolver symbol table
